hello-world/requirements.txt
dnspython==2.3.0
nslookup==1.7.0
